services/presentation/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
neo4j==5.15.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy-file==0.2.0

# langchain==0.1.20
# langchain-core==0.1.28
# langchain-community==0.0.38
# langchain-text-splitters==0.0.1
# langchain-ollama==0.1.0
# ollama==0.3.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-ollama==0.1.3
ollama==0.3.1
openai==1.30.0

python-multipart==0.0.6
python-pptx==0.6.23
reportlab==4.0.7
Pillow==10.1.0
aiofiles==23.2.1